Player and Game Screen Prototype
I added basic barriers for design concept, the ship cannot fly through
the barriers as of now but collision detection could be improved I
believe. I also added a single bullet fire and a basic game menu. I
tried to comment as much as I could to what I added so that it was at
least easy to follow.