First functional prototype of game
Game is almost to a par-finished state. Asteroids randomly generate,
bullet detection is implemented, points are awarded for hits, player is
bounded properly (but can be glitched if you try). Minor bits are still
unfinished however: namely the game still does not end unless the user
chooses to end it.