Added Database connection and Sound Effects
Working game with sound effects and Database connection. At the moment
the user can choose to add or not add their score to the High Score DB;
either way, at the end of each play through the game will print out the
contents of the High Score DB to the console. The JDBC .jar file has
been added to the project and should work if your computer has MySQL
setup and the proper database/table imported to your localhost (I've
added the exported DB to the project in 'Database backup' so you can
import it).